+++ src/doctor.rs
@@ -1,12 +1,17 @@
-use std::{collections::BTreeMap, process::Command};
+use std::{
+    collections::{BTreeMap, HashMap},
+    process::Command,
+};
 
+use colored::Colorize;
+use lazy_static::lazy_static;
 use regex::Regex;
 use semver::Version;
 
 use crate::{
     bgworker,
     config::{self, Config, Database},
-    db, Error, Result,
+    db, depcheck, Error, Result,
 };
 
 const SEAORM_INSTALLED: &str = "SeaORM CLI is installed";
@@ -19,12 +24,28 @@ const QUEUE_CONN_OK: &str = "queue connection: success";
 const QUEUE_CONN_FAILED: &str = "queue connection: failed";
 const QUEUE_NOT_CONFIGURED: &str = "queue not configured?";
 
+// versions health
+const MIN_SEAORMCLI_VER: &str = "1.1.0";
+lazy_static! {
+    static ref MIN_DEP_VERSIONS: HashMap<&'static str, &'static str> = {
+        let mut min_vers = HashMap::new();
+
+        min_vers.insert("tokio", "1.33.0");
+        min_vers.insert("sea-orm", "1.1.0");
+        min_vers.insert("validator", "0.18.0");
+        min_vers.insert("axum", "0.7.5");
+
+        min_vers
+    };
+}
+
 /// Represents different resources that can be checked.
 #[derive(PartialOrd, PartialEq, Eq, Ord, Debug)]
 pub enum Resource {
     SeaOrmCLI,
     Database,
-    Redis,
+    Queue,
+    Deps,
 }
 
 /// Represents the status of a resource check.
@@ -93,17 +114,58 @@ impl std::fmt::Display for Check {
 /// Runs checks for all configured resources.
 /// # Errors
 /// Error when one of the checks fail
-pub async fn run_all(config: &Config) -> Result<BTreeMap<Resource, Check>> {
-    let mut checks = BTreeMap::from([
-        (Resource::SeaOrmCLI, check_seaorm_cli()?),
-        (Resource::Database, check_db(&config.database).await),
-    ]);
+pub async fn run_all(config: &Config, production: bool) -> Result<BTreeMap<Resource, Check>> {
+    let mut checks = BTreeMap::from([(Resource::Database, check_db(&config.database).await)]);
 
     if config.workers.mode == config::WorkerMode::BackgroundQueue {
-        checks.insert(Resource::Redis, check_queue(config).await);
+        checks.insert(Resource::Queue, check_queue(config).await);
+    }
+
+    if !production {
+        checks.insert(Resource::Deps, check_deps()?);
+        checks.insert(Resource::SeaOrmCLI, check_seaorm_cli()?);
     }
 
     Ok(checks)
+}
+
+/// Checks "blessed" / major dependencies in a Loco app Cargo.toml, and
+/// recommend to update.
+/// Only if a dep exists, we check it against a min version
+/// # Errors
+/// Returns error if fails
+pub fn check_deps() -> Result<Check> {
+    let cargolock = fs_err::read_to_string("Cargo.lock")?;
+
+    let crate_statuses = depcheck::check_crate_versions(&cargolock, MIN_DEP_VERSIONS.clone())?;
+    let mut report = String::new();
+    report.push_str("Dependencies\n");
+    let mut all_ok = true;
+
+    for status in &crate_statuses {
+        if let depcheck::VersionStatus::Invalid {
+            version,
+            min_version,
+        } = &status.status
+        {
+            report.push_str(&format!(
+                "     {}: version {} does not meet minimum version {}\n",
+                status.crate_name.yellow(),
+                version.red(),
+                min_version.green()
+            ));
+            all_ok = false;
+        }
+    }
+    Ok(Check {
+        status: if all_ok {
+            CheckStatus::Ok
+        } else {
+            CheckStatus::NotOk
+        },
+        message: report,
+        description: None,
+    })
 }
 
 /// Checks the database connection.
@@ -160,7 +222,6 @@ pub async fn check_queue(config: &Config) -> Check {
     }
 }
 
-const MIN_SEAORMCLI_VER: &str = "1.1.0";
 /// Checks the presence and version of `SeaORM` CLI.
 /// # Panics
 /// On illegal regex