+++ src/lib.rs
@@ -6,6 +6,7 @@ pub use self::errors::Error;
 
 mod banner;
 pub mod bgworker;
+mod depcheck;
 pub mod initializers;
 pub mod prelude;